requirements/runtime.txt
matplotlib
mmcv>=0.3.1
numpy
# need older pillow until torchvision is fixed
Pillow<=6.2.2
six
terminaltables
torch>=1.1
torchvision
tqdm
seaborn
